Major changes in packaging, changed existsByUsername into existByName and added it to AttorneyRepository.
