+++ src/main/java/businessrule/gateway/AttorneyGateway.java
@@ -1,7 +1,5 @@
 package businessrule.gateway;
 
-import driver.database.UserGateway;
-
 public interface AttorneyGateway extends UserGateway {
 
 }
