+++ src/main/java/businessrule/gateway/ClientGateway.java
@@ -1,9 +1,5 @@
 package businessrule.gateway;
 
-import driver.database.UserGateway;
-
 public interface ClientGateway extends UserGateway {
-
-    boolean existsByUsername(String username);
 
 }
