+++ src/main/java/businessrule/usecase/SelectQuestionInteractor.java
@@ -7,7 +7,7 @@ import driver.database.QuestionGateway;
 import businessrule.responsemodel.TheQuestionResponseModel;
 import businessrule.outputboundary.TheQuestionOutputBoundary;
 import entity.User;
-import driver.database.UserGateway;
+import businessrule.gateway.UserGateway;
 import businessrule.gateway.UserGatewayFactory;
 import java.util.Map;
