+++ src/main/java/businessrule/usecase/ViewQuestionInteractor.java
@@ -3,7 +3,7 @@ package businessrule.usecase;
 import businessrule.inputboundary.ViewInputBoundary;
 import businessrule.requestmodel.ViewRequestModel;
 import driver.database.QuestionGateway;
-import driver.database.UserGateway;
+import businessrule.gateway.UserGateway;
 import businessrule.gateway.UserGatewayFactory;
 import businessrule.outputboundary.ViewOutputBoundary;
 import businessrule.responsemodel.ViewResponseModel;
